Number game!
This is an old ruby file I made in June of this year. I had known Ruby for three days, and I was so excited that I posted this to the /r/ruby subreddit (amazing people over there, seriously go check it out!). This is an incredibly rough file (it works, but it looks horrendous) but it's a great milestone for me. This is when I first started learning! I want to keep it. For growth-showing reasons. Yay!