Much better version of Guessing Game.
This is the much-refined version of the guessing game, the code was provided by Reddit user 0x0dea. I didn't want to post it to anything until I understood what was going on in the code (in June I didn't, but now I do!). Another milestone in helping me get enthused to code! Woo!