Let's make a list!
This program makes a list of items you wish to add to it. You can add as many as you want, and they can be any kind of item. The program asks for your item, item quantity and if you want to add more items to the list. Choosing "No" will break the loop, and it will generate a list for you on screen. I tried to comment in the code and explain what was going on, for demonstration sake. I'm really proud of this program, this is the most advanced thing in Ruby I've done so far and it took me about 15-30 minutes to actually create. It doesn't look great, by any means, but I like it. It does the job.